+++ src/app/page.tsx
@@ -42,7 +42,7 @@ const SCHEDULE_ALIASES: Record<string, string> = {
 type SavedTeam = { id: string; name: string; logo?: string; savedAt: string }
 
 function fmtKickoff(iso: string) {
-  return new Date(iso).toLocaleTimeString('en-US', { hour: 'numeric', minute: '2-digit', timeZoneName: 'short' })
+  return new Date(iso).toLocaleTimeString('en-US', { hour: 'numeric', minute: '2-digit', timeZoneName: 'long' })
 }
 function fmtDate(iso: string) {
   return new Date(iso).toLocaleDateString('en-US', { weekday: 'long', month: 'long', day: 'numeric' })
